GaussianMixtureRegressions

Documentation for GaussianMixtureRegressions.

GaussianMixtureRegressions.AbstractGMR

GaussianMixtureRegressions.aic

GaussianMixtureRegressions.bic

GaussianMixtureRegressions.membership

GaussianMixtureRegressions.nll

StatsAPI.loglikelihood

GaussianMixtureRegressions.AbstractGMR — Type

AbstractGMR

Abstract type for Gaussian mixture regression.

GaussianMixtureRegressions.aic — Method

aic(model[, X, y])

Akaike information criterion of a mixture model.

Returns Akaike information criterion values. If X, y are not given, AIC is evaluated by training data.

GaussianMixtureRegressions.bic — Method

bic(model[, X, y])

Bayesian information criterion of a mixture model.

Returns Bayesian information criterion values. If X, y are not given, BIC is evaluated by training data.

GaussianMixtureRegressions.membership — Method

membership(model::GMR[, k_component], X)

Return the probability of membership for an observation.

GaussianMixtureRegressions.nll — Method

nll(model, X, y)

Negative log likelihood for a model.

Returns negative log likelihood for a linear regression model, which are evaluated by data.

StatsAPI.loglikelihood — Method

loglikelihood(model[, X, y])

Log likelihood of a model.

Returns log likelihood values. If X, y are not given, log likelihood is evaluated by training data.